+++ core/listen.py
@@ -8,7 +8,7 @@ import subprocess
 import time
 from watchdog.observers import Observer
 from watchdog.events import FileSystemEventHandler
-from utils import load_config
+from utils.shortcuts import load_config
 
 config = load_config()
 dev = config.get('dev')
