+++ core/utils/exception.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+
+class InvalidPage(Exception):
+    pass
